+++ src/components/display/index.ts
@@ -2,6 +2,7 @@ export { FormFooter, type FormFooterProps } from './formFooter';
 export { FormWrapper, type FormWrapperProps } from './formWrapper';
 export { LightboxImage, type LightboxImageProps } from './lightboxImage';
 export { SetupGuide, type SetupGuideProps } from './setupGuide';
+export { CodeSnippet, type CodeSnippetProps } from './codeSnippet';
 export { GuideComponent } from './guide';
 export { RenderGuide, type RenderGuideProps } from './setupGuide/section';
 export { Link, type LinkProps } from './link';
