+++ src/components/display/setupGuide/section/index.tsx
@@ -2,7 +2,7 @@ import { PluginUiMessageHandler } from '@cloudquery/plugin-config-ui-connector';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 
-import { CodeSnippet } from './codeSnippet';
+import { CodeSnippet } from '../../codeSnippet';
 import { ConditionalRenderingWrapper } from '../../../controls/conditionalRenderingWrapper';
 import { LightboxImage } from '../../lightboxImage';
